requirement.txt
lightning
torchmetrics
wandb
tqdm
hydra-core
matplotlib
pandas
plotly
ray[default]
scikit-learn
scipy
